requirements.txt
colorama
google-generativeai
requests
